% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_wrappers.R
\name{run_grid_both}
\alias{run_grid_both}
\title{Run Illumina + PacBio grid simulation}
\usage{
run_grid_both(simref_fa, tag)
}
\arguments{
\item{simref_fa}{Character. Reference FASTA.}

\item{tag}{Character. Tag used for output subfolders.}
}
\value{
TRUE on success.
}
\description{
Simulates read sets across fixed Illumina and PacBio coverage grids.
}
\examples{
\dontrun{
run_grid_both("01_simref/ecoli_repMed.fa", "ecoli_repMed")
}
}
