% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_wrappers.R
\name{run_quast_grid}
\alias{run_quast_grid}
\title{Run QUAST evaluation across assembly grid}
\usage{
run_quast_grid(tag)
}
\arguments{
\item{tag}{Character. Tag used for input/output folders.}
}
\value{
TRUE on success.
}
\description{
Run QUAST evaluation across assembly grid
}
\examples{
\dontrun{
run_quast_grid("ecoli_repMed")
}
}
