% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_wrappers.R
\name{run_unicycler_grid}
\alias{run_unicycler_grid}
\title{Run Unicycler hybrid assemblies across a grid}
\usage{
run_unicycler_grid(tag, threads = 8)
}
\arguments{
\item{tag}{Character. Tag used for input/output folders.}

\item{threads}{Integer. Number of threads.}
}
\value{
TRUE on success.
}
\description{
Run Unicycler hybrid assemblies across a grid
}
\examples{
\dontrun{
run_unicycler_grid("ecoli_repMed", threads = 8)
}
}
